+++ demos/meridian/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @substrat-run/demo-hr
 
+## 0.5.8
+
+### Patch Changes
+
+- Updated dependencies [f869541]
+- Updated dependencies [f869541]
+- Updated dependencies [19fb697]
+- Updated dependencies [f869541]
+- Updated dependencies [717600e]
+- Updated dependencies [46b1cac]
+- Updated dependencies [9208b4e]
+  - @substrat-run/kernel@0.72.0
+  - @substrat-run/adapter-sqlite@0.72.0
+  - @substrat-run/adapter-cloudflare@0.72.0
+  - @substrat-run/contracts@0.72.0
+  - @substrat-run/vertical-host@0.72.0
+  - @substrat-run/engine-protocol@0.8.0
+  - @substrat-run/control-plane-api@0.72.0
+  - @substrat-run/connector-scrive@0.10.0
+  - @substrat-run/engine-absence@0.2.3
+
 ## 0.5.7
 
 ### Patch Changes