+++ demos/rally/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @substrat-run/demo-rally
 
+## 0.0.69
+
+### Patch Changes
+
+- Updated dependencies [f869541]
+- Updated dependencies [f869541]
+- Updated dependencies [19fb697]
+- Updated dependencies [f869541]
+- Updated dependencies [717600e]
+- Updated dependencies [46b1cac]
+- Updated dependencies [9208b4e]
+  - @substrat-run/kernel@0.72.0
+  - @substrat-run/adapter-sqlite@0.72.0
+  - @substrat-run/contracts@0.72.0
+  - @substrat-run/engine-booking@0.2.3
+  - @substrat-run/engine-invites@0.2.3
+  - @substrat-run/engine-invoicing@0.6.3
+
 ## 0.0.68
 
 ### Patch Changes